:bug: [bug]: 로그인 이후 뒤로가기 버튼에서 발생하는 에러 해결
로그인이 된 이후 브라우저의 뒤로가기버튼을 클릭하면 다시 로그인이 진행되는 문제를 해결 합니다
