+++ src/pages/index.ts
@@ -5,3 +5,4 @@ export * from './main';
 export * from './submit';
 export * from './status';
 export * from './login';
+export * from './loginprocess';
