+++ src/pages/loginprocess/LoginProcessPage.tsx
@@ -1,0 +1,12 @@
+import { LoginProcess } from '@/components';
+import React from 'react';
+
+const LoginProcessPage = () => {
+  return (
+    <div>
+      <LoginProcess />
+    </div>
+  );
+};
+
+export default LoginProcessPage;
