+++ src/pages/loginprocess/index.ts
@@ -1,0 +1,1 @@
+export { default as LoginProcessPage } from './LoginProcessPage';
